requirements.txt
requests
beautifulsoup4
click
dnspython
argparse
